:bug: [maykinmedia/open-klant#414] fix help text not showing
:green_heart: [maykinmedia/open-klant#414] ci fix
